+++ lib/InvertedClipper.dart
@@ -14,9 +14,9 @@ class InvertedClipper extends CustomClipper<Path> {
         if (widgetData.isEnabled) {
           final GlobalKey key = widgetData.key;
           if (key == null) {
-            throw new Exception("GlobalKey is null!");
+        //    throw new Exception("GlobalKey is null!");
           } else if (key.currentWidget == null) {
-            throw new Exception("GlobalKey is not assigned to a Widget!");
+//            throw new Exception("GlobalKey is not assigned to a Widget!");
           } else {
             areas.add(getHoleArea(key: key,shape: widgetData.shape,padding: widgetData.padding));
           }